refactor: RequestMessageHandler로 이름 변경
